requirements.txt
langchain_openai
langchain_core
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
chromadb
faiss-cpu
groq
cassio
beautifulsoup4
langchain-groq
wikipedia
arxiv
langchainhub
sentence_transformers
PyPDF2
langchain-objectbox
datasets
scipy
torch
transformers
langchain
flask
pandas
matplotlib
yfinance
numpy
pycoingecko
scikit-learn
tensorflow 
numpy>=1.23.0
pandas>=1.5.0
scikit-learn>=1.1.0
requests>=2.28.0
Pillow>=9.0.0
matplotlib>=3.5.0
